Implement v0 prototype.
Support a subset of /meta RESTful API, see README
for details. The code is versioning-ready, it relies
on blueprints. Build and dependencies are properl
implemented. Properl mysql parameter binding are used
for security reasons. It implements a small "metaBackend"
script for loading metadata information. It contains
a small reinit.py test program.